backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
mysqlclient==2.2.8
gunicorn==23.0.0
whitenoise==6.8.2
Pillow==11.1.0
sqlparse==0.5.5
asgiref==3.11.1
django-ckeditor-5==0.2.2
